csyUIandSound02

+++ ProjectB/Assets/Script/Player/Fire_HardCandy.cs
@@ -10,7 +10,7 @@ public class Fire_HardCandy : MonoBehaviour
     Rigidbody2D rb;
     SpriteRenderer spriteRenderer;
     public float timer = 1f;
-    public float damage = 10f;
+    public int damage = 10;
     [SerializeField] Sprite[] hardCandySprites;
 
     public void SetDirection()
